Android Fundamentals 1.2B.HW
Added a Zero Button which resets the counter.  The Zero Button is grey (inactive) unless the count is greater than zero.

The Count button also changes color to purple when an odd number is displayed.

One possible bug is that the Count button does NOT change colors when the count is reset.  So it is possible for the count to be 0 and the button to be purple.  But Hey, zero is neither odd nor even so I guess its technically not incorrect right?